Adding descriptive classes
